refresh priv to pub

+++ src/bondgame.h
@@ -2,8 +2,8 @@
 #define _BOND_H_
 #include "ultra64.h"
 
-extern u16 cfb_16_a[];
-extern u16 cfb_16_b[];
+extern u8 cfb_16_a[];
+extern u8 cfb_16_b[];
 
 extern u32 *_codeSegmentStart;
 extern u32 *_codeSegmentEnd;
